requirements.txt
clickhouse-driver>=0.2.0
PyYAML>=6.0
